Improve sns linac benchmark
